requirements.txt
certifi==2023.7.22
chardet==3.0.4
googletrans==3.1.0a0
greenlet==2.0.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pydantic==1.10.12
PySimpleGUI==4.60.5
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
typing_extensions==4.7.1
python-decouple==3.8
